Moved ApacheDS functionality into LdapUtility
